rebuild of quantum engine for use of USTRUCT as opposed to UCLASS
Given UObject derived classes were not meant to be invoked or constructed directly in code I have gone the route of using it as the topmost layer for anything that does not involve actors or actor components.  USTRUCT's still maintain most of the C++ functionality while using the UE4 TArray and TMap tools.  Successful compilation so now the debugging of the new setup starts.