tested gate matrix creation, did code cleanup and adding UE4 interface
Successful creation of matrix/operator gates, moving on to testing if matrix/operator mulitplication works as intended. Further code cleanup and testing.  Attempting now to create UE4 Interface class to act as template for other implementations.  Simulating a quantum computer as fast as possible can be a tricky business when you are varying which gates you use and how many (controlled gates decide the level of entanglement). Therefore it is important that there is polymorphism to allow future implementations.